% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{clueless}
\alias{clueless}
\title{simplest detective}
\usage{
clueless(data, ...)
}
\arguments{
\item{data}{data frame, typically coming from [utils::getParseData()]}

\item{...}{ignored}
}
\description{
simplest detective
}
