% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_file.R
\name{css_file}
\alias{css_file}
\title{helper function to get a css file}
\usage{
css_file(filename = "default.css")
}
\arguments{
\item{filename}{name of file}
}
\value{
the name of the first file that is found
}
\description{
helper function to get a css file
}
\details{
the search goes as follows: 
- absolutely or relative to the working directory
- then the directory ~/.R/highlight
- then the stylesheet directory in the installed package
}
