% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{document_header}
\alias{document_header}
\alias{document_footer}
\title{header and footer for full document}
\usage{
document_header(css = "default.css")

document_footer()
}
\arguments{
\item{css}{css file, see [css_file()]}
}
\description{
header and footer for full document
}
