% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/renderer.R
\name{formatter_html}
\alias{formatter_html}
\title{html formatter}
\usage{
formatter_html(tokens, styles, ...)
}
\arguments{
\item{tokens}{tokens to wrap}

\item{styles}{styles to give to the tokens}

\item{\dots}{ignored}
}
\description{
Wraps tokens into span tags with the class corresponding to the style
}
\seealso{
\code{\link{renderer_html}}
}

