% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{formatter_latex}
\alias{formatter_latex}
\title{Latex formatter}
\usage{
formatter_latex(tokens, styles, ...)
}
\arguments{
\item{tokens}{vector of okens}

\item{styles}{vector of styles}

\item{\dots}{ignored}
}
\value{
A vector of latex commands
}
\description{
Combines tokens and styles into a latex command
}
\examples{
formatter_latex( "hello world", "blue" )
}
\seealso{
\code{\link{renderer_latex}}
}

