% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{header_latex}
\alias{footer_latex}
\alias{header_latex}
\title{latex header and footer}
\usage{
header_latex(document, styles, boxes, minipage = FALSE)

footer_latex(document, minipage = FALSE)
}
\arguments{
\item{document}{logical. If TRUE the header and footer functions will create the 
full document (including preamble with boxes and styles)}

\item{styles}{a vector of style definitions to include in the preamble if document is TRUE}

\item{boxes}{a vector of boxes definitions to include in the preamble if document is TRUE}

\item{minipage}{if \code{TRUE}, the highlighted latex is included in a minipage environment}
}
\value{
A function is returned, suitable for the header or footer argument
of the latex renderer
}
\description{
These functions return appropriate header and footer functions
for the latex renderer
}
\examples{
  
h <- header_latex( document = FALSE )
h()
f <- footer_latex( document = FALSE )
f()
}

