\name{highlight}
\Rdversion{1.1}
\alias{highlight}
\title{
	syntax highlighting
}
\description{
	syntax highlighting
}
\usage{
highlight(x, detective, renderer)
}
\arguments{
  \item{x}{
what to highlight
}
  \item{detective}{
the detective finds evidence about the syntax
}
  \item{renderer}{
the renderer uses the evidence to render the syntax into a markup format
}
}
\value{
	the rendered text
}
\author{
	Romain Francois <francoisromain@free.fr>
}
\examples{
	# TODO
}
\keyword{ manip }
