\name{highlight_themes}
\alias{highlight_themes}
\title{
List of themes supported by external_highlight
}
\description{
List of themes supported by external_highlight
}
\usage{
highlight_themes()
}
\value{
A character vector with the names of the themes
}
\examples{
highlight_themes()  
}
\keyword{manip}

