% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{hl_hook_source}
\alias{hl_hook_source}
\alias{hl_hook_css}
\title{knitr hooks}
\usage{
hl_hook_source(x, options)

hl_hook_css(x)
}
\arguments{
\item{x}{the code}

\item{options}{chunk options}
}
\description{
knitr hooks
}
