% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{moriarty}
\alias{moriarty}
\title{moriarty}
\usage{
moriarty(assistant = lestrade, char = "▪️")
}
\arguments{
\item{assistant}{initial detective

This starts by the investigation of the assistant, then replaces tokens
by a character.}
}
\description{
moriarty
}
