% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{mycroft}
\alias{mycroft}
\title{mycroft}
\usage{
mycroft(assistant = lestrade)
}
\arguments{
\item{assistant}{initial detective

This starts by the investigation of the assistant, then adds the css class "focus" to 
tokens depending on the `packages` and `functions` options.}
}
\description{
mycroft
}
