% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer_latex.R
\name{space_latex}
\alias{newline_latex}
\alias{renderer_latex}
\alias{space_latex}
\title{LaTeX renderer}
\usage{
space_latex()

newline_latex()

renderer_latex(document = TRUE, stylesheet = "default", minipage = FALSE)
}
\arguments{
\item{document}{logical. Should the renderer create the full document or only the code
section, assuming the document is already created. Using FALSE 
is used by the sweave driver shipped with this package.}

\item{stylesheet}{stylesheet to use.}

\item{minipage}{if TRUE, the highlighted latex is included in a minipage environment}

\item{boxes}{a function that returns definitions of latex boxes used for non standard
characters. The reason for using boxes is that some character need 
to be escaped to be rendered, and unfortunately, escaping turns
alltt off, which does not produce satisfying rendering. This argument
is used by the header function when the document argument is TRUE. 
It is also used in the sweave driver at the very beginning of the document}

\item{styles}{style definitions inferred from the parsing of the stylesheet. See \code{\link{styler}} and
\code{\link{styler_assistant_latex}}.}
}
\value{
a \samp{renderer} object, suitable for the \samp{renderer} argument of 
\code{\link{highlight}}.
}
\description{
renderer implementation targetting latex markup. The result
markup uses the latex \samp{alltt} package to achieve true type 
renderering and therefore does not depend on verbatim-like environments.
}
\examples{
\dontrun{
	r <- renderer_latex(document = T )
	r$space()
	r$newline()
	r$boxes()
	r$translator( "# the hash symbol gets a latex box" )
}
}

