% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweaveLatexDriver.R
\name{runcode_external}
\alias{runcode_external}
\title{runcode with external highlight (andre version)}
\usage{
runcode_external(object, chunk, options)
}
\description{
runcode with external highlight (andre version)
}

