% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweaveLatexDriver.R
\name{runcode_weave}
\alias{runcode_weave}
\title{run the code by weaving (similar to standard driver)}
\usage{
runcode_weave(object, chunk, options)
}
\description{
run the code by weaving (similar to standard driver)
}

