% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{sherlock}
\alias{sherlock}
\title{Sherlock Holmes, highlighting detective}
\usage{
sherlock(assistant = lestrade)
}
\arguments{
\item{assistant}{initial detective}
}
\description{
Sherlock Holmes, highlighting detective
}
