% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/renderer.R
\name{styler_assistant_latex}
\alias{styler_assistant_latex}
\title{latex styler assistant}
\usage{
styler_assistant_latex(x)
}
\arguments{
\item{x}{output from \code{\link{css.parser}}}
}
\value{
a vector of latex style definitions corresponding to (a subset of) the
        output of the parser
}
\description{
This function takes the output of the \code{\link{css.parser}} and
produces latex style definitions from it.
}
\details{
The function create a new latex command for each css declaration, i.e.
each item of the list \samp{x} it is passed.

The assistant currently honours the following css settings: color,
\samp{text-decoration:underline}, \samp{font-weight:bold[er]} and
\samp{font-style:italic}
}
\seealso{
\code{\link{styler}}
}

