% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer_latex.R
\name{translator_latex}
\alias{translator_latex}
\title{LaTeX translator}
\usage{
translator_latex(x, size = c("normalsize", "tiny", "scriptsize",
  "footnotesize", "small", "large", "Large", "LARGE", "huge", "Huge"))
}
\arguments{
\item{x}{text to translate}

\item{size}{font size}
}
\value{
translated text
}
\description{
This function translates character vectors so that they nicely print
in LaTeX. In particular this uses latex boxes.
}
\seealso{
the latex renderer: \code{\link{renderer_latex}} uses this translator.
}

