% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{getStyleFile}
\alias{getStyleFile}
\title{helper function to get a style file}
\usage{
getStyleFile(name = "default", extension = "css")
}
\arguments{
\item{name}{the name of the style file to look for}

\item{extension}{the file extension (css, sty, or xterm)}
}
\value{
the name of the first file that is found, or NULL
}
\description{
helper function to get a style file
}
\details{
the search goes as follows: first the current working directory
then the directory ~/.R/highlight, then the stylesheet directory
in the installed package
}
