% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\docType{package}
\name{highlight-package}
\alias{highlight-package}
\title{Syntax Highlighter for R}
\description{
Syntax highlighter for R based on output from the R parser
}
\examples{
\dontrun{
highlight( "rnorm(10)" )

}
}
\seealso{
The main function of the package is \code{\link{highlight}}.
}
