\name{highlight_output_types}
\alias{highlight_output_types}
\title{
List of available output types supported by external_highlight
}
\description{
    List of available output types supported by \code{\link{external_highlight}}
}
\usage{
highlight_output_types()
}
\value{
A character vector with the list of supported types
}
\seealso{
    \code{\link{external_highlight}}
}
\examples{
    highlight_output_types()
}
\keyword{manip}
