% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_highlight.R
\name{highlight_output_types}
\alias{highlight_output_types}
\title{List of available output types supported by external_highlight}
\usage{
highlight_output_types()
}
\value{
A character vector with the list of supported types
}
\description{
List of available output types supported by \code{\link{external_highlight}}
}
