% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_highlight.R
\name{highlight_themes}
\alias{highlight_themes}
\title{List of themes supported by external_highlight}
\usage{
highlight_themes()
}
\value{
A character vector with the names of the themes
}
\description{
List of themes supported by \code{\link{external_highlight}}
}
