% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer_html.R
\name{renderer_html}
\alias{renderer_html}
\title{html renderer using span tags and CSS}
\usage{
renderer_html(document = TRUE, stylesheet = "default")
}
\arguments{
\item{document}{logical. Indicates if the renderer should render a full document
or simply a \samp{<pre>} section containing the highlighted
tokens.}

\item{stylesheet}{stylesheet to use. This is used by the header when document is TRUE.
The content of the stylesheet is copied verbatim into a \samp{<style>}
tag in that case. See \code{\link{getStyleFile}} for details
on where the stylesheet can be located}
}
\value{
A renderer suitable for the \samp{renderer} argument of \code{\link{highlight}}
}
\description{
implementation of the \code{\link{renderer}} that renders
the information as a series of \samp{<span>} html tags
}
\seealso{
\code{\link{renderer}} for a description of the interface
	this renderer is implementing. 
	
	\code{\link{highlight}} takes a renderer argument to which it delegates rendering.
}
