% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer_latex.R
\name{renderer_latex}
\alias{renderer_latex}
\title{LaTeX renderer}
\usage{
renderer_latex(document = TRUE, stylesheet = "default", minipage = FALSE)
}
\arguments{
\item{document}{logical. Should the renderer create the full document or only the code
section, assuming the document is already created. Using FALSE 
is used by the sweave driver shipped with this package.}

\item{stylesheet}{stylesheet to use.}

\item{minipage}{if TRUE, the highlighted latex is included in a minipage environment}

\item{styles}{style definitions inferred from the parsing of the stylesheet. See \code{\link{styler}} and
\code{\link{styler_assistant_latex}}.}
}
\value{
a \samp{renderer} object, suitable for the \samp{renderer} argument of 
\code{\link{highlight}}.
}
\description{
renderer implementation targetting latex markup. The result
markup uses the latex \samp{alltt} package to achieve true type 
renderering and therefore does not depend on verbatim-like environments.
}
\examples{
\dontrun{
	r <- renderer_latex(document = T )
	r$space()
	r$newline()
	r$boxes()
	r$translator( "# the hash symbol gets a latex box" )
}
}
