% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{sherlock}
\alias{sherlock}
\title{Sherlock Holmes, highlighting detective}
\usage{
sherlock(data)
}
\arguments{
\item{data}{data frame, typically coming from [utils::getParseData()]}
}
\description{
Sherlock Holmes, highlighting detective
}
