% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{sherlock}
\alias{sherlock}
\title{Sherlock Holmes, highlighting detective}
\usage{
sherlock(data, palette = muted_colors, ...)
}
\arguments{
\item{data}{data frame, typically coming from [utils::getParseData()]}

\item{palette}{a function converting numbers from 0 to 1 into a color}
}
\description{
Sherlock Holmes, highlighting detective
}
