% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectives.R
\name{simple_detective}
\alias{simple_detective}
\title{Simple detective}
\usage{
simple_detective(x, ...)
}
\arguments{
\item{x}{output of the parser. The detective is only interested in the 
\samp{token} column of the data.}

\item{\dots}{ignored}
}
\value{
a vector of styles grouping similar tokens together
}
\description{
This detective only uses semantic information to make its 
investigation.
}
\examples{
\dontrun{
p <- parse( text = deparse( jitter ), keep.source=TRUE )
simple_detective( p )
}
}
